+++ source/Internal/Framework/Module/Configuration/DataObject/ModuleConfiguration.php
@@ -15,8 +15,10 @@ use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration\SmartyPluginDirectory;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration\Template;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration\TemplateBlock;
+use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\Exception\InvalidModuleIdException;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\Exception\ModuleSettingNotFountException;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Setting\Setting;
+use function Symfony\Component\String\u;
 
 class ModuleConfiguration
 {
@@ -116,9 +118,14 @@ class ModuleConfiguration
      * @param string $id
      *
      * @return ModuleConfiguration
+     * @throws InvalidModuleIdException
      */
     public function setId(string $id): ModuleConfiguration
     {
+        if (u($id)->containsAny('/')) {
+            throw new InvalidModuleIdException('Module ID (' . $id . ') must not contain "/".');
+        }
+
         $this->id = $id;
 
         return $this;