+++ tests/Integration/Internal/Framework/Module/Configuration/Dao/ModuleTemplateExtensionChainTest.php
@@ -27,11 +27,11 @@ final class ModuleTemplateExtensionChainTest extends TestCase
         parent::setUp();
 
         $this->shopConfigurationDao = $this->getContainer()->get(ShopConfigurationDaoInterface::class);
+
     }
 
     public function testGetWithNonExistingTemplate(): void
     {
-        $this->markTestSkipped('TODO: fix');
         $nonExistingTemplate = uniqid('template_', true);
         $moduleTemplateExtensions = $this->shopConfigurationDao
             ->get($this->shopId)